Responses API - tools, tool calls, tool outputs, inputs, settings, json formats

+++ openai-core/src/main/scala/io/cequence/openaiscala/domain/responsesapi/tools/ComputerUseTool.scala
@@ -4,15 +4,15 @@ package io.cequence.openaiscala.domain.responsesapi.tools
  * A tool that controls a virtual computer.
  *
  * @param displayHeight
- *   The height of the computer display.
+ *   The height of the computer display (in pixels).
  * @param displayWidth
- *   The width of the computer display.
+ *   The width of the computer display (in pixels).
  * @param environment
  *   The type of computer environment to control.
  */
 case class ComputerUseTool(
-  displayHeight: Double,
-  displayWidth: Double,
+  displayHeight: Int,
+  displayWidth: Int,
   environment: String
 ) extends Tool {
   val `type`: String = "computer_use_preview"
